changed all scripts with corrected cvMD functions in sensors, source, blocks and 20s in timeg 👽

+++ archive_/rsa_blocks_net_old.py
@@ -63,28 +63,50 @@ def process_subject(subject, network, epoch_num):
     gc.collect()
         
     blocks = np.unique(behav.blocks)
+    all_pats, all_rands = [], []
 
-    for block in blocks:
-        
-        block = int(block)
-        print(f"Processing {subject} - session {epoch_num} - block {block} - {network}")
+    if not op.exists(res_dir / f"pat-{epoch_num}.npy") or not op.exists(res_dir / f"rand-{epoch_num}.npy") or overwrite:
 
-        if not op.exists(res_dir / f"pat-{epoch_num}-{block}.npy") or overwrite:
-            filter = (behav["trialtypes"] == 1) & (behav["blocks"] == block)     
-            X_pat = stcs_data[filter]
-            y_pat = behav[filter].reset_index(drop=True).positions
-            assert len(X_pat) == len(y_pat)
-            rdm_pat = loocv_mahalanobis_parallel(X_pat, y_pat, 1)
-            np.save(res_dir / f"pat-{epoch_num}-{block}.npy", rdm_pat)
-        
-        if not op.exists(res_dir / f"rand-{epoch_num}-{block}.npy") or overwrite:
-            filter = (behav["trialtypes"] == 2) & (behav["blocks"] == block)            
-            X_rand = stcs_data[filter]
-            y_rand = behav[filter].reset_index(drop=True).positions
-            assert len(X_rand) == len(y_rand)
-            rdm_rand = loocv_mahalanobis_parallel(X_rand, y_rand, 1)
-            np.save(res_dir / f"rand-{epoch_num}-{block}.npy", rdm_rand)                
+        for block in blocks:
+            
+            block = int(block)
+            print(f"Processing {subject} - session {epoch_num} - block {block} - {network}")
+
+            if not op.exists(res_dir / f"pat-{epoch_num}-{block}.npy") or overwrite:
+                filter = (behav["trialtypes"] == 1) & (behav["blocks"] == block)     
+                X_pat = stcs_data[filter]
+                y_pat = behav[filter].reset_index(drop=True).positions
+                assert len(X_pat) == len(y_pat)
+                rdm_pat = loocv_mahalanobis_fixed(X_pat, y_pat)
+                np.save(res_dir / f"pat-{epoch_num}-{block}.npy", rdm_pat)
+            else:
+                rdm_pat = np.load(res_dir / f"pat-{epoch_num}-{block}.npy")
+            all_pats.append(rdm_pat)
+            
+            if not op.exists(res_dir / f"rand-{epoch_num}-{block}.npy") or overwrite:
+                filter = (behav["trialtypes"] == 2) & (behav["blocks"] == block)            
+                X_rand = stcs_data[filter]
+                y_rand = behav[filter].reset_index(drop=True).positions
+                assert len(X_rand) == len(y_rand)
+                rdm_rand = loocv_mahalanobis_fixed(X_rand, y_rand)
+                np.save(res_dir / f"rand-{epoch_num}-{block}.npy", rdm_rand)
+            else:
+                rdm_rand = np.load(res_dir / f"rand-{epoch_num}-{block}.npy")
+            all_rands.append(rdm_rand)
                 
+        all_pats, nan_pat = interpolate_rdm_nan(np.array(all_pats))
+        if nan_pat:
+            print(subject, "has pattern nans interpolated in session", epoch_num, network)
+        all_rands, nan_rand = interpolate_rdm_nan(np.array(all_rands))
+        if nan_rand:
+            print(subject, "has random nans interpolated in session", epoch_num, network)
+        
+        np.save(res_dir / f"pat-{epoch_num}.npy", all_pats)
+        np.save(res_dir / f"rand-{epoch_num}.npy", all_rands)
+    
+    else:
+        print("Files already exist, skipping", subject, epoch_num, network)
+        
     del stcs_data
     gc.collect()
